pyproject.toml
[project]
name = "reelforge"
version = "0.1.0"
description = "The modular video creation platform - Forge your perfect reel engine"
authors = [
    {name = "ReelForge Team"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0.0",
    "pydantic>=2.0.0",
    "loguru>=0.7.0",
    "pyyaml>=6.0.0",
    "edge-tts>=7.2.3",
    "certifi>=2025.10.5",
    "comfykit>=0.1.0",
    "ffmpeg-python>=0.2.0",
    "httpx>=0.28.1",
    "pillow>=10.0.0,<12",
    "html2image>=2.0.7",
    "streamlit>=1.40.0",
    "openai>=2.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.6.0",
]

[project.scripts]
reelforge = "reelforge.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

